Move APIErrorResponse to HTTPError
